fix(freeze): complete Decimal128 hardening + cross-language dataset_root (streaming)
The post-fix re-review proved the first Decimal128 fix was INCOMPLETE — 5 more verified bugs the
conformance gate (recode-only, main-encoder-only) couldn't reach. All fixed + newly gated:
- Go: Decimal128.Scale int8 -> int32 (was silently truncating canonical |scale|>127, diverging from
  Rust); EncodeToWriter scatter encoder scale raw-byte -> svarint; lowest-terms now rejects coeff==0&&
  scale!=0; json.go formatDecimal128 reads coefficient little-endian (was big-endian -> wrong JSON).
  New golden_scatter_test gates the scatter path (EncodeToWriter==golden for all 73).
- TS: DeterministicEncoder scale raw-byte -> svarint; lowest-terms coeff==0 reject; new
  decimal_sorted_parity.test gates the sorted encoder vs the main encoder.
- Rust: verified already-correct (stores raw svarint, rejects coeff==0&&scale!=0) — no fix needed.
- Coverage added so this class can't hide again: golden decimal_scale_200 / _neg200 (|scale|>127),
  negative decimal_zero_scale (coeff0,scale!=0 -> ERR_NON_CANONICAL). Floors -> golden 73, neg 26.

Streaming (Phase 3) cross-language complete: recode --dataset-root in Go/Rust/TS (Merkle DAG over the
ordered shard file-roots, reusing the gated §7 merkle); NEW dataset_identity_gate.py, all 4 agree (4/4).

Independently verified: full 11-gate suite GREEN (count-guard · pytest · conformance 73 · content-addr 73
· file 7/7 · tensor 8/8 · fingerprint 73/73 · DATASET 4/4 · negative lenient+strict 26 exact codes · fuzz)
+ native suites green (Go, Rust 10 suites, TS 370). Lesson: alternate encoders + JSON bridge were
untested paths — now gated.

+++ go/new.go
@@ -28,7 +28,7 @@ func Float64(f float64) *Value {
 }
 
 // NewDecimal128 returns a decimal128 value.
-func NewDecimal128(scale int8, coef [16]byte) *Value {
+func NewDecimal128(scale int32, coef [16]byte) *Value {
 	return &Value{typ: TypeDecimal128, decimal128: Decimal128{Scale: scale, Coef: coef}}
 }
