+++ go/types.go
@@ -203,8 +203,8 @@ func (t Type) String() string {
 // Decimal128 represents a 128-bit decimal number.
 // Value = Coef * 10^(-Scale)
 type Decimal128 struct {
-	Scale int8     // -127 to +127
-	Coef  [16]byte // Two's complement big-endian
+	Scale int32    // SPEC scale is an int64 svarint; int32 covers the decimal domain (e.g. |scale|>127)
+	Coef  [16]byte // Two's complement little-endian (matches the §2.3 wire form)
 }
 
 // ============================================================
